requirements.txt
# --- Core ML ---
torch==2.0.1
numpy==1.26.4
sentence-transformers==3.3.1

# --- Web framework ---
fastapi==0.115.5
uvicorn[standard]==0.30.6

# --- LangChain & Tavily ---
langchain==0.3.27
langchain-core==0.3.77
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langsmith==0.4.31
tavily-python==0.3.3

# --- Utilities ---
pydantic==2.11.9
orjson==3.11.3
PyYAML==6.0.3
SQLAlchemy==2.0.43
tenacity==9.1.2
requests==2.32.5
httpx==0.28.1
